requirements.txt
aioredis==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.5.1
Automat==20.2.0
billiard==4.2.0
CacheControl==0.13.1
cachetools==5.3.2
celery==5.4.0
certifi==2022.6.15
cffi==1.15.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cron_descriptor==1.4.0
cryptography==37.0.2
daphne==4.0.0
Deprecated==1.2.13
Django==4.2
django-celery-beat==2.5.0
django-ckeditor==6.0.0
django-cors-headers==3.13.0
django-debug-toolbar==4.2.0
django-filter==22.1
django-friendship==1.9.6
django-jazzmin==3.0.0
django-js-asset==2.1.0
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-api-key==3.0.0
dnspython==2.6.1
firebase-admin==6.2.0
google-api-core==2.12.0
google-api-python-client==2.106.0
google-auth==2.29.0
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.0
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.2
grpcio-status==1.59.2
gunicorn==20.1.0
hiredis==2.3.2
httplib2==0.22.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.4
incremental==21.3.0
kombu==5.3.7
livekit_server_sdk_python @ git+https://github.com/livekit/livekit-server-sdk-python.git@0ad70a1823a74822702029581a50c2e73c8625e6
Markdown==3.3.7
msgpack==1.0.4
packaging==21.3
pillow==11.3.0
prompt-toolkit==3.0.29
proto-plus==1.22.3
protobuf==4.21.12
psycopg2-binary==2.9.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pymediainfo==5.1.0
pymongo==4.7.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.6
python-ffmpeg-video-streaming==0.1.15
python-resize-image==1.1.20
pytz==2022.1
redis==4.5.4
requests==2.28.1
rsa==4.9
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
structlog==22.3.0
tinify==1.5.2
twirp==0.0.7
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.2.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.12
vine==5.1.0
wcwidth==0.2.5
websocket-client==1.8.0
whitenoise==6.4.0
wrapt==1.14.1
zipp==3.8.0
zope.interface==5.4.0
